version: fix for shifted GET_VERSION_API in r3p1
We abuse the fact that GET_API_VERSION is the last _MALI_UK_CORE_SUBSYSTEM
ioctl, and that it goes from nr 5 to 3 in r3p1. So in r3p1, calling ioctl 5
results in -EPERM, which we can nicely catch and then we can call ioctl 3
to get the proper version.

Also add support for detecting r3p2.

Signed-off-by: Luc Verhaegen <libv@skynet.be>